requirements.txt
appdirs==1.4.3
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
autopep8==1.4.4
cached-property==1.5.1
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0
django-filter==2.2.0
djangorestframework==3.10.3
gunicorn==20.0.4
idna==2.8
isodate==0.6.0
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.4.2
Markdown==3.1.1
mccabe==0.6.1
phonenumberslite==8.11.1
psycopg2==2.8.4
pycodestyle==2.5.0
pyhubtel-sms==0.0.1
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
requests-toolbelt==0.9.1
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.7
wrapt==1.11.2
zeep==3.4.0
